1.0.2
Bug fixed, ban retrieval system changed

+++ Modules/on_join.py
@@ -24,38 +24,39 @@ class Events(commands.Cog):
     @commands.Cog.listener()
     async def on_member_join(self, member):
         bot = self.bot
-        count = 0
-        reasons = []
-        config = await Configer.get(Configer, member.guild.id, "modchannel")
+        config = await Configer.get(self, member.guild.id, "modchannel")
         configid = int(config)
-        for guild in bot.guilds:
-            try:
-                ban = await guild.fetch_ban(member)
-                reasons.append(f"\n {guild}: {ban.reason}")
-                count += 1
-            except discord.NotFound:
-                pass
-            except Exception as e:
-                print(e)
-        sr = "".join(reasons)
+        bans = bot.bans
+        if f"{member.id}" in bans:
+            print("member in bans")
+            reasons = []
+            count = 0
+            for guild in bot.guilds:
+                try:
+                    ban = bans[f"{member.id}"][f"{guild.id}"]['reason']
+                    reasons.append(f"\n {guild}: {ban}")
+                    count += 1
+                except discord.NotFound:
+                    pass
+                except Exception as e:
+                    pass
+            sr = "".join(reasons)
 
-        if count >= 1:
-
-            if configid == 0:
-                channel = member.guild.system_channel
-                await channel.send("Moderation channel is not set, use **/config change**")
-            else:
-                channel = bot.get_channel(configid)
-                if len(sr) < 1800:
-                    await channel.send(f"{member.mention} is banned in: {sr}")
+            if count >= 1:
+                if configid == 0:
+                    channel = member.guild.system_channel
+                    await channel.send("Moderation channel is not set, use **/config change**")
                 else:
-                    with open(f"{random.randint(1, 1000)}.txt", 'w', encoding='utf-8') as f:
-                        f.write(sr)
-                    await channel.send(f"{member.mention} is banned in:", file=discord.File(f.name, "banned.txt"))
-
-                    os.remove(f.name)
+                    channel = bot.get_channel(configid)
+                    if len(sr) < 1800:
+                        await channel.send(f"{member.mention} is banned in: {sr}")
+                    else:
+                        with open(f"{random.randint(1, 1000)}.txt", 'w', encoding='utf-8') as f:
+                            f.write(sr)
+                        await channel.send(f"{member.mention} is banned in:", file=discord.File(f.name, "banned.txt"))
+                        os.remove(f.name)
         else:
-            print("user not banned, skipping")
+            print(f"{member.guild}: {member} has no ban record")
 
 
 async def setup(bot):
